+++ src/widgets/todo/ui/index.ts
@@ -1,1 +1,1 @@
-export * from './todo';
+export * from './TodoCard';
